Setup food detail by food id.

+++ app/src/main/java/com/adedom/myfood/presentation/component/Screen.kt
@@ -28,7 +28,7 @@ sealed class Screen(
             "foodDetail/{foodId}",
             listOf(navArgument("foodId") { type = NavType.IntType }),
         ) {
-            fun arguments(foodId: Int): String {
+            fun arguments(foodId: Long): String {
                 return "foodDetail/$foodId"
             }
         }
